requirement.txt
asgiref==3.3.1
boto3==1.17.9
botocore==1.20.9
Django==3.1.6
djangorestframework==3.12.2
jmespath==0.10.0
python-dateutil==2.8.1
python-ffmpeg-video-streaming==0.1.5
pytz==2021.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
gunicorn
